+++ 7월/MySummary2/app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "30.0.1"
 
     defaultConfig {
-        applicationId "com.example.summary"
+        applicationId "com.example.summary2"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
