+++ 7월/MySummary2/app/src/androidTest/java/com/example/summary2/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.summary;
+package com.example.summary2;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.summary", appContext.getPackageName());
+        assertEquals("com.example.summary2", appContext.getPackageName());
     }
 }
