+++ 7월/MySummary2/app/src/main/java/com/example/summary2/CommentItemView.java
@@ -1,0 +1,32 @@
+package com.example.summary2;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.LayoutInflater;
+import android.widget.LinearLayout;
+
+import androidx.annotation.Nullable;
+
+public class CommentItemView extends LinearLayout {
+
+
+    public CommentItemView(Context context) {
+        super(context);
+
+        init(context);
+    }
+
+    public CommentItemView(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+
+        init(context);
+    }
+
+    private void init(Context context){
+        // 인플레이션을 통한 xml 파일과의 연결
+        LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        inflater.inflate(R.layout.comment_item_view, this, true);
+
+
+    }
+}
