--- 7월/MySummary2/app/src/main/res/drawable/linear_border_purple.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <stroke android:width="2dp" android:color="#800080"/>
-    <solid android:color="#00000000"/>
-
-</shape>
