+++ 7월/MySummary2/app/src/main/res/drawable/thumb_up_selector.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_pressed="true"
+        android:drawable="@drawable/ic_thumb_up_selected"/>
+
+    <item android:drawable="@drawable/ic_thumb_up"/>
+
+</selector>
