--- 7월/MySummary2/app/src/main/res/drawable/thumbnail_base.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <gradient
-        android:startColor="#7288DB"
-        android:centerColor="#3250B4"
-        android:endColor="#254095"
-        android:angle="90"
-        android:centerY="0.5"
-        />
-    <corners android:radius="2dp"/>
-</shape>
