+++ 7월/MySummary2/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">씨네마천국</string>
+    <string name="app_name">MySummary2</string>
 </resources>
