+++ 7월/MySummary2/app/src/test/java/com/example/summary2/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.summary;
+package com.example.summary2;
 
 import org.junit.Test;
