+++ 7월/MySummary2/settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "MySummary"
+rootProject.name = "MySummary2"
